+++ cards_api/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class CardsConfig(AppConfig):
-    name = 'cards'
+    name = 'cards_api'
