+++ cards_api/urls.py
@@ -1,0 +1,7 @@
+from django.conf.urls import url
+
+from . import views
+
+urlpatterns = [
+    url(r'^', views.CardsListJsonView.as_view()),
+]
