+++ python_ru/settings.py
@@ -15,7 +15,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
 
-    'cards',
+    'cards_api',
 ]
 
 MIDDLEWARE = [
